+++ CHANGELOG.md
@@ -15,6 +15,7 @@ au](https://solidcommunity.au/docs/solidui)
 
 ## 0.4.0 Refine and Tune
 
++ Adds additional layout width checks [0.3.6 20260326 tonypioneer]
 + Add checkbox and new webid widgets to SolidLogin() [0.3.5 20260325 tonypioneer]
 + Add WebID to SetupWizard [0.3.4 20260325 tonypioneer]
 + Login again if webid changed [0.3.3 20260319 tonypioneer]